Added new functionalities

+++ fitness_world/photos/views.py
@@ -1,13 +1,19 @@
+from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.shortcuts import render, redirect
 from django.urls import reverse_lazy
 from django.views import generic
 
+from fitness_world import settings
 from fitness_world.common.forms import CommentCreateForm
+from fitness_world.common.views import page_not_found
+from fitness_world.core.core_utils import user_permissions
 from fitness_world.core.photo_utils import photo_likes_count, photo_is_liked_by_user
 from fitness_world.photos.forms import CreatePhotoForm, EditPhotoForm, DeletePhotoForm
 from fitness_world.photos.models import Photo
+
+UserModel = get_user_model()
 
 
 class CreatePhotoView(LoginRequiredMixin, generic.CreateView):
@@ -46,10 +52,25 @@ class DetailsPhotoView(LoginRequiredMixin, generic.DetailView):
         return context
 
 
-class EditPhotoView(LoginRequiredMixin, generic.UpdateView):
+class EditPhotoView(LoginRequiredMixin, UserPassesTestMixin, generic.UpdateView):
     template_name = 'photos/photo-edit-page.html'
     form_class = EditPhotoForm
     model = Photo
+
+    def test_func(self):
+        try:
+            photo = Photo.objects.get(pk=self.kwargs.get('pk'))
+            user_pk = photo.user.pk
+            owner = UserModel.objects.filter(pk=user_pk).get()
+            return user_permissions(self.request, owner)
+        except Photo.DoesNotExist:
+            page_not_found(self.request)
+
+    def handle_no_permission(self):
+        if not self.request.user.is_authenticated:
+            return redirect(settings.LOGIN_URL)
+
+        return page_not_found(self.request)
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
@@ -68,6 +89,9 @@ class EditPhotoView(LoginRequiredMixin, generic.UpdateView):
 def delete_photo(request, pk):
     photo = Photo.objects.filter(pk=pk).get()
 
+    if not user_permissions(request, photo.user):
+        return page_not_found(request)
+
     if request.method == 'GET':
         form = DeletePhotoForm(instance=photo)
     else:
@@ -83,4 +107,3 @@ def delete_photo(request, pk):
     }
 
     return render(request, 'photos/photo-delete-page.html', context)
-
